Split operations module into submodules
